Fabric_assigment/sdk/network/channel_manager.go
package main

import (
	"fmt"

	"github.com/hyperledger/fabric-sdk-go/pkg/common/providers/fabric"
	"github.com/hyperledger/fabric-sdk-go/pkg/fabsdk"
)

type ChannelManager struct{}

func (cm *ChannelManager) CreateChannel(ctx fabsdk.TransactionContext, channelName string) error {
	fmt.Println("Creating channel...")
	// Implement channel creation logic here
	return nil
}

func (cm *ChannelManager) JoinChannel(ctx fabsdk.TransactionContext, channelName string) error {
	fmt.Println("Joining channel...")
	// Implement channel joining logic here
	return nil
}

func (cm *ChannelManager) GetChannel(ctx fabsdk.TransactionContext, channelName string) ([]byte, error) {
	fmt.Println("Getting channel...")
	// Implement channel retrieval logic here
	return nil, nil
}